requirements.txt
Pillow>=10.0.0
Flask>=3.0.0
tqdm>=4.65.0
tkinterdnd2>=0.3.0
